pyproject.toml
[project]
name = "my-debauch-devotion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "python-telegram-bot[job-queue]>=22.1",
    "pytz>=2025.2",
    "requests>=2.32.3",
]
